+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+
+-   Fix `{Array,List}.find_map`
+
 ## 0.2.0
 
 -   Add gen_js_api conversion function for `error` type